implement date and boolean filters
