+++ packages/www/hooks/use-api.tsx
@@ -557,6 +557,7 @@ const makeContext = (state: ApiState, setState) => {
         limit: string;
         cursor: string;
         order: string;
+        active?: string;
       }
     ): Promise<[Stream[], string]> {
       const filters = opts?.filters ? JSON.stringify(opts?.filters) : undefined;
@@ -564,6 +565,7 @@ const makeContext = (state: ApiState, setState) => {
         `/stream?${qs.stringify({
           userId,
           filters,
+          active: opts?.active,
           order: opts?.order,
           limit: opts?.limit,
           cursor: opts?.cursor,
